Fix tricore UB (#2204)

+++ bindings/python/capstone/tricore.py
@@ -7,14 +7,14 @@ from .tricore_const import *
 class TriCoreOpMem(ctypes.Structure):
     _fields_ = (
         ('base', ctypes.c_uint8),
-        ('disp', ctypes.c_int32),
+        ('disp', ctypes.c_int64),
     )
 
 
 class TriCoreOpValue(ctypes.Union):
     _fields_ = (
         ('reg', ctypes.c_uint),
-        ('imm', ctypes.c_int32),
+        ('imm', ctypes.c_int64),
         ('mem', TriCoreOpMem),
     )
